完善项目功能：添加社区论坛、点赞评论、头像优化等功能，并生成完整文档

+++ app/src/main/java/com/example/icyclist/adapter/ForumCategoryAdapter.kt
@@ -45,7 +45,7 @@ class ForumCategoryAdapter(private val categories: List<ForumCategory>) :
             val context = holder.itemView.context
             val intent = Intent(context, TopicListActivity::class.java).apply {
                 putExtra("CATEGORY_NAME", category.name)
-                putExtra("CATEGORY_ID", category.id.toLong())
+                putExtra("CATEGORY_ID", category.id)
             }
             context.startActivity(intent)
         }
