+++ app/src/main/java/com/example/icyclist/adapter/TopicAdapter.kt
@@ -41,7 +41,7 @@ class TopicAdapter(private val topics: List<Topic>) :
             val context = holder.itemView.context
             val intent = Intent(context, TopicDetailActivity::class.java).apply {
                 putExtra("TOPIC_TITLE", topic.title)
-                putExtra("TOPIC_ID", topic.id.toLong())
+                putExtra("TOPIC_ID", topic.id)
             }
             context.startActivity(intent)
         }
